rename electrodes to ephys
